pyproject.toml
[project]
name = "rag-chrome-plugin"
version = "0.1.0"
description = "RAG Chrome Plugin - Intelligent webpage indexing and semantic search with agentic architecture"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.4.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.2",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.3",
    "markitdown>=0.0.1a2",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = [
    "*.py",
    "faiss_cache/*.json",
    "faiss_cache/*.bin"
]
